refactor: MBTI별 통계 페이지 폴더 구조 변경
