+++ src/pages/stats/statsMbti/StatsMbti.styles.ts
