+++ src/pages/stats/statsMbti/StatsMbti.tsx
